requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.28.0
python-dotenv>=1.0.0
python-multipart>=0.0.5
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
pydantic>=2.0.0 